plugins/source/typeform/requirements.txt
cloudquery-plugin-sdk==0.0.7
pyarrow==12.0.1
requests==2.31.0
pytest==7.4.0
pandas==2.0.3